requirements.txt
fastapi
uvicorn
sentence-transformers
torch
Pillow
